WIP: Implement unit types with type safety
- Add Satoshi and Dollar wrapper types with explicit conversions
- Implement Equals and GetHashCode for Satoshi comparison operators
- Move JSON converters to separate Converters namespace
- Prevent integer division truncation in financial calculations
- Update models and services to use typed units
- Fix compilation errors with explicit conversions

+++ tests/AutoBot.Tests/Services/ProcessTradeExecutionTests.cs
@@ -575,11 +575,11 @@ public class ProcessTradeExecutionTests
         decimal exitPrice,
         decimal leverage,
         TradeSide side,
-        decimal expectedMargin,
+        long expectedMargin,
         decimal expectedLiquidation,
-        decimal expectedOpeningFee,
-        decimal expectedClosingFee,
-        decimal expectedPL)
+        long expectedOpeningFee,
+        long expectedClosingFee,
+        long expectedPL)
     {
         // Act
         var trade = TradeFactory.CreateTrade(
